+++ tests_integration/test_data/projects/snowpark_glob_patterns/app_2/d.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+from snowflake.snowpark import Session
+
+
+# test import
+import syrupy
+
+
+def test_procedure(session: Session) -> str:
+    return "Test procedure"